refactor(test): move reconciler-layer tests from service module to reconciler module
StatsPriorityClassTest, ReconcileExecutionPolicyTest, ReconcileCapturePolicyTest,
and InMemoryStorePriorityDispatchTest were placed in service/src/test because the
reconciler module had pre-existing compile errors at the time they were written.
The reconciler module now compiles cleanly, so all four tests are relocated to
their correct home alongside the classes they test.
