+++ reconciler/src/test/java/ai/floedb/floecat/reconciler/jobs/ReconcileCapturePolicyTest.java
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package ai.floedb.floecat.service.reconciler.jobs;
+package ai.floedb.floecat.reconciler.jobs;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import ai.floedb.floecat.reconciler.jobs.ReconcileCapturePolicy;
 import ai.floedb.floecat.stats.spi.JobCostHint;
 import java.util.List;
 import java.util.Set;
