+++ reconciler/src/test/java/ai/floedb/floecat/reconciler/jobs/impl/InMemoryStorePriorityDispatchTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package ai.floedb.floecat.service.reconciler.jobs;
+package ai.floedb.floecat.reconciler.jobs.impl;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -26,7 +26,6 @@ import ai.floedb.floecat.reconciler.jobs.ReconcileJobKind;
 import ai.floedb.floecat.reconciler.jobs.ReconcileJobStore;
 import ai.floedb.floecat.reconciler.jobs.ReconcileScope;
 import ai.floedb.floecat.reconciler.jobs.StatsPriorityClass;
-import ai.floedb.floecat.reconciler.jobs.impl.InMemoryReconcileJobStore;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
